--- a/generate.go
+++ b/generate.go
@@ -18,8 +18,6 @@ import (
 // * rhs (rdata)
 // But we are lazy here, only the range is parsed *all* occurrences
 // of $ after that are interpreted.
-// Any error are returned as a string value, the empty string signals
-// "no error".
 func (zp *ZoneParser) generate(l lex) (RR, bool) {
 	token := l.token
 	step := 1