This is a pretty large pull request, but again it's ultimately rather simple. The `parseZone` goroutine is eliminated and replaced with a new `ZoneParser` type with a simple chan-like API. This eliminates the goroutines and the 10,000 element buffered `parseZone` chan, which will substantially reduce memory usage.

Errors such as #786 should be impossible with the new API.

`ParseZone` maintains the same semantics as previous, including using a buffered chan of 10,000 elements. It is now a wrapper around the `ZoneParser` API.

The API (minus the `SetX` methods) was discussed in https://github.com/miekg/dns/issues/786#issuecomment-429630595.

**Note:** This pull request contains an API change. `NewRR` and `ReadRR` no longer support the `$INCLUDE` directive. See the `SetIncludeAllowed` comment for why.

This is the comparison between tip (17c1bc6792fdf1918200e9a675cdf2e3c9d585cd) and the pull request:
```
name          old time/op    new time/op    delta
NewRR-12        6.57µs ± 2%    2.42µs ± 2%  -63.19%  (p=0.000 n=24+24)
ReadRR-12       6.99µs ± 1%    2.77µs ± 5%  -60.36%  (p=0.000 n=25+24)
ParseZone-12    33.9µs ±11%    40.3µs ± 9%  +18.93%  (p=0.000 n=25+23)

name          old alloc/op   new alloc/op   delta
NewRR-12          808B ± 0%      816B ± 0%   +0.99%  (p=0.000 n=25+25)
ReadRR-12       1.82kB ± 0%    1.82kB ± 0%   +0.44%  (p=0.000 n=25+25)
ParseZone-12    83.9kB ± 0%    84.0kB ± 0%   +0.19%  (p=0.000 n=25+25)

name          old allocs/op  new allocs/op  delta
NewRR-12          17.0 ± 0%      15.0 ± 0%  -11.76%  (p=0.000 n=25+25)
ReadRR-12         19.0 ± 0%      17.0 ± 0%  -10.53%  (p=0.000 n=25+25)
ParseZone-12      91.0 ± 0%      92.0 ± 0%   +1.10%  (p=0.000 n=25+25)
```

This is the comparison between `ParseZone` (old) and `ZoneParser` (new)—both after this pull request:
```
name            old time/op    new time/op    delta
benchmark-12    40.3µs ± 9%    10.6µs ± 1%  -73.80%  (p=0.000 n=23+24)

name            old alloc/op   new alloc/op   delta
benchmark-12    84.0kB ± 0%     1.6kB ± 0%  -98.10%  (p=0.000 n=25+25)

name            old allocs/op  new allocs/op  delta
benchmark-12      92.0 ± 0%      81.0 ± 0%  -11.96%  (p=0.000 n=25+25)
```

Fixes #786
Updates coredns/coredns#2194