@miekg 2018-10-15
[ Quoting <notifications@github.com> in "[miekg/dns] Add new ZoneParser API ..." ]
>This is a pretty large pull request, but again it's ultimately rather simple. The `parseZone` goroutine is eliminated and replaced with a new `ZoneParser` type with a simple chan-like API. This eliminates the goroutines and the 10,000 element buffered `parseZone` chan, which will substantially reduce memory usage.
>
>Errors such as #786 should be impossible with the new API.
>
>`ParseZone` maintains the same semantics as previous, including using a buffered chan of 10,000 elements. It is now a wrapper around the `ZoneParser` API.
>
>The API (minus the `SetX` methods) was discussed in https://github.com/miekg/dns/issues/786#issuecomment-429630595.
>
>**Note:** This pull request contains an API change. `NewRR` and `ReadRR` no longer support the `$INCLUDE` directive. See the `SetIncludeAllowed` comment for why.

This is awesome, but why the API change? Even if it's a bad idea your code will 
break subtlely (instead of not compiling). Can we reverse it, by setting to 
allow by default?