@tmthrgd 2018-10-16
This is what the coredns diff would look like:
```diff
diff --git a/plugin/file/file.go b/plugin/file/file.go
index 2711282ebc..4b01a3fa61 100644
--- a/plugin/file/file.go
+++ b/plugin/file/file.go
@@ -120,33 +120,31 @@ func (s *serialErr) Error() string {
 // If serial >= 0 it will reload the zone, if the SOA hasn't changed
 // it returns an error indicating nothing was read.
 func Parse(f io.Reader, origin, fileName string, serial int64) (*Zone, error) {
-	tokens := dns.ParseZone(f, dns.Fqdn(origin), fileName)
-	defer func() {
-		// Drain the tokens chan so that large zone files won't
-		// leak goroutines and memory.
-		for range tokens {
-		}
-	}()
+	zp := dns.NewZoneParser(f, dns.Fqdn(origin), fileName)
+	zp.SetIncludeAllowed(true)
+
 	z := NewZone(origin, fileName)
-	seenSOA := false
-	for x := range tokens {
-		if x.Error != nil {
-			return nil, x.Error
-		}
 
-		if !seenSOA && serial >= 0 {
-			if s, ok := x.RR.(*dns.SOA); ok {
-				if s.Serial == uint32(serial) { // same serial
+	seenSOA := false
+	for rr, ok := zp.Next(); ok; rr, ok = zp.Next() {
+		if !seenSOA {
+			if s, ok := rr.(*dns.SOA); ok {
+				if serial >= 0 && s.Serial == uint32(serial) { // same serial
 					return nil, &serialErr{err: "no change in SOA serial", origin: origin, zone: fileName, serial: serial}
 				}
 				seenSOA = true
 			}
 		}
 
-		if err := z.Insert(x.RR); err != nil {
+		if err := z.Insert(rr); err != nil {
 			return nil, err
 		}
 	}
+
+	if err := zp.Err(); err != nil {
+		return nil, err
+	}
+
 	if !seenSOA {
 		return nil, fmt.Errorf("file %q has no SOA record", fileName)
 	}
```