@miekg 2018-10-17
more tests and they pass as do the old tests. I think this is good to go - hard to properly review due to the amount of changes, trusting the tests a little here.

btw: I think the reader bufsize for NewRR can be even smaller, like 256