@tmthrgd 2018-10-17
I’ll wait for your reply on the review comments above. I tried to make it easy to review, but I accept it’s a large PR.

`NewRR` isn’t buffered at all because `strings.Reader` implements `io.ByteReader` which hits the fast path in `newZLexer`.