@tmthrgd 2018-10-18
> Ok, that's fine, but can you add a comment to this, so its more clear?

@miekg Will do.

> How to use the ZoneParser API, mostly?

@miekg Isn't that covered by the documentation of `ZoneParser`? It includes example code equivalent to the example code on `ParseZone`.